fashion-runway-analysis/requirements.txt
# Fashion Runway Analysis - Dépendances Python

# Core Data Science
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.3.0
torch>=2.0.0
torchvision>=0.15.0

# Transformers & Vision Models
transformers>=4.30.0
pillow>=10.0.0

# Image Processing
opencv-python>=4.8.0
scikit-image>=0.21.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress bars
tqdm>=4.65.0

# Configuration & Utils
pyyaml>=6.0
pathlib>=1.0.1 # Normalement intégré mais au cas où
